Split interfaces from customizer TS
