+++ src/ts/admin/LiveReloadComputedProperty.ts
@@ -1,0 +1,4 @@
+interface LiveReloadComputedProperty {
+  additionalSettings?: Array<string>;
+  value: (value: any, additionalValues: Array<any>) => string;
+}
