+++ src/ts/admin/LiveReloadMediaRules.ts
@@ -1,0 +1,4 @@
+interface LiveReloadMediaRules {
+  minWidth?: number;
+  maxWidth?: number;
+}
