+++ src/ts/admin/LiveReloadProperty.ts
@@ -1,0 +1,6 @@
+interface LiveReloadProperty {
+  name: string;
+  prefix?: string;
+  postfix?: string;
+  computed?: LiveReloadComputedProperty;
+}
