+++ src/ts/admin/LiveReloadTarget.ts
@@ -1,0 +1,5 @@
+interface LiveReloadTarget {
+  selector: string;
+  media?: LiveReloadMediaRules;
+  properties: Array<LiveReloadProperty>;
+}
